+++ Cartrekk/Cartrekk/LocationTrackingService.swift
@@ -72,40 +72,70 @@ class LocationTrackingService: NSObject, ObservableObject, CLLocationManagerDele
     }
     
     // MARK: - Route Data Management
-    func saveRoute(raw_userId: String, time: TimeInterval) -> Route {
-        let route = Route(
-            id: UUID(),
-            date: Date(),
-            coordinates: locations.map {
-                RouteCoordinate(
-                    latitude: $0.coordinate.latitude,
-                    longitude: $0.coordinate.longitude,
-                    timestamp: $0.timestamp
-                )
-            }
-        )
-        let routeId = route.id.uuidString
-        let distance = totalDistance * 0.00062137
-        let duration = time
-        let likes = 0
-        let polyline = Polyline(locations: locations)
-        let encodedPolyline: String = polyline.encodedPolyline
-        let isPublic = true
-        let routeImages: [String]? = nil
-        let userId = raw_userId
+    func saveRoute(raw_userId: String, time: TimeInterval, routeID: UUID) {
+        let routeId = routeID.uuidString
         
-        // Here you would typically save to persistent storage
-        FirestoreManager.shared.saveRouteDetails(routeId: routeId,
-                         distance: distance,
-                         duration: duration,
-                         likes: likes,
-                         polyline: encodedPolyline,
-                         isPublic: isPublic,
-                         routeImages: routeImages,
-                         userId: userId)
-        return route
+        // First get the existing document to preserve routeImages
+        FirestoreManager.shared.db.collection("routes").document(routeId).getDocument { [weak self] (document, error) in
+            guard let self = self else { return }
+            
+            let existingImages = document?.data()?["routeImages"] as? [String] ?? []
+            
+            let distance = self.totalDistance * 0.00062137
+            let duration = time
+            let likes = 0
+            let polyline = Polyline(locations: self.locations)
+            let encodedPolyline: String = polyline.encodedPolyline
+            let isPublic = false
+            let userId = raw_userId
+            
+            // Save to Firebase with existing images
+            FirestoreManager.shared.saveRouteDetails(
+                routeId: routeId,
+                distance: distance,
+                duration: duration,
+                likes: likes,
+                polyline: encodedPolyline,
+                isPublic: isPublic,
+                routeImages: existingImages,  // Use the existing images
+                userId: userId
+            )
+        }
+    }
+    
+    func initialize_route(routeID: UUID?){
+        let route_id = routeID!.uuidString
+        FirestoreManager.shared.saveRouteDetails(routeId: route_id, distance: 0, duration: 0.0, likes: 0, polyline: "emptyRoute", isPublic: false, routeImages: nil, userId: "emptyUser")
+    }
+    
+    func addImageToRoute(routeID: UUID?, imageURL: String) {
+        guard let route_id = routeID?.uuidString else {
+            print("Invalid route ID")
+            return
+        }
+        
+        let routeRef = FirestoreManager.shared.db.collection("routes").document(route_id)
+        
+        routeRef.getDocument { (document, error) in
+            if let error = error {
+                print("Error getting document: \(error)")
+                return
+            }
+            
+            guard let document = document, document.exists,
+                  var data = document.data() else {
+                print("Document does not exist or is empty")
+                return
+            }
+            
+            // Get current routeImages array or create empty one
+            var currentImages = data["routeImages"] as? [String] ?? []
+            
+            // Append new image URL
+            currentImages.append(imageURL)
+            
+            FirestoreManager.shared.saveRouteDetails(routeId: route_id, distance: 0, duration: 0.0, likes: 0, polyline: "emptyRoute", isPublic: false, routeImages: currentImages, userId: "emptyUser")
+            
+        }
     }
 }
-
-// MARK: - Data Models
-
