--- Cartrekk/Cartrekk/UserProfile.swift
@@ -1,10 +1,0 @@
-//
-//  UserProfile.swift
-//  Cartrekk
-//
-//  Created by Tejas Vaze on 2/12/25.
-//
-
-
-
-
